+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 12.2
 - Use multiple Select2 instead of checkboxes in PortalPollsNotes.fnc.php
 - Simplify code to save PUBLISHED_PROFILES in PortalNotes.php & PortalPolls.php
 - Use multiple Select2 instead of checkboxes in Resources.php & Resources.fnc.php
+- Enrollment End Date: No N/A option if next entry already has Start date in StudentsUsersInfo.fnc.php
 
 Changes in 12.1.3
 -----------------